+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - default `cairo-coverage` command. You must now use `cairo-coverage run` to generate a coverage report.
 - support for windows. We recommend using the Windows Subsystem for Linux (WSL) to run `cairo-coverage` on Windows.
+- support for scarb `2.8.*`.
+- option to include test functions in the coverage report as this only worked for `2.8.*`. It Will be brought back in
+  the future.
 
 ## [0.5.0] - 2025-03-06